Created Vm_Guest_Memory trait and changed MmapMemory accordingly
