+++ src/linux/uhyve.rs
@@ -4,10 +4,9 @@
 use std::{
 	cmp,
 	ffi::OsString,
-	fmt, mem,
-	os::raw::c_void,
+	fmt,
 	path::{Path, PathBuf},
-	ptr::{self, NonNull},
+	ptr,
 	sync::{Arc, Mutex},
 };
 
@@ -15,19 +14,13 @@ use hermit_entry::boot_info::RawBootInfo;
 use kvm_bindings::*;
 use kvm_ioctls::VmFd;
 use log::debug;
-use nix::sys::mman::*;
 use vmm_sys_util::eventfd::EventFd;
-use x86_64::{
-	structures::paging::{Page, PageTable, PageTableFlags, Size2MiB},
-	PhysAddr,
-};
 
 use crate::{
 	consts::*,
-	linux::{virtio::*, x86_64::kvm_cpu::KvmCpu, KVM},
+	linux::{mem::MmapMemory, virtio::*, x86_64::kvm_cpu::KvmCpu, KVM},
 	params::Params,
-	vm::Vm,
-	x86_64::create_gdt_entry,
+	vm::{Vm, VmGuestMemory},
 	HypervisorResult,
 };
 
@@ -163,7 +156,7 @@ impl Uhyve {
 			"gdbstub is only supported with one CPU"
 		);
 
-		let hyve = Uhyve {
+		let mut hyve = Uhyve {
 			vm,
 			offset: 0,
 			entry_point: 0,
@@ -241,50 +234,8 @@ impl Vm for Uhyve {
 	}
 
 	/// Initialize the page tables for the guest
-	fn init_guest_mem(&self) {
-		debug!("Initialize guest memory");
-
-		let (mem_addr, _) = self.guest_mem();
-
-		unsafe {
-			let pml4 = &mut *((mem_addr as u64 + BOOT_PML4.as_u64()) as *mut PageTable);
-			let pdpte = &mut *((mem_addr as u64 + BOOT_PDPTE.as_u64()) as *mut PageTable);
-			let pde = &mut *((mem_addr as u64 + BOOT_PDE.as_u64()) as *mut PageTable);
-			let gdt_entry: u64 = mem_addr as u64 + BOOT_GDT.as_u64();
-
-			// initialize GDT
-			*((gdt_entry) as *mut u64) = create_gdt_entry(0, 0, 0);
-			*((gdt_entry + mem::size_of::<*mut u64>() as u64) as *mut u64) =
-				create_gdt_entry(0xA09B, 0, 0xFFFFF); /* code */
-			*((gdt_entry + 2 * mem::size_of::<*mut u64>() as u64) as *mut u64) =
-				create_gdt_entry(0xC093, 0, 0xFFFFF); /* data */
-
-			/* For simplicity we currently use 2MB pages and only a single
-			PML4/PDPTE/PDE. */
-
-			// per default is the memory zeroed, which we allocate by the system call mmap
-			/*libc::memset(pml4 as *mut _ as *mut libc::c_void, 0x00, PAGE_SIZE);
-			libc::memset(pdpte as *mut _ as *mut libc::c_void, 0x00, PAGE_SIZE);
-			libc::memset(pde as *mut _ as *mut libc::c_void, 0x00, PAGE_SIZE);*/
-
-			pml4[0].set_addr(
-				BOOT_PDPTE,
-				PageTableFlags::PRESENT | PageTableFlags::WRITABLE,
-			);
-			pml4[511].set_addr(
-				BOOT_PML4,
-				PageTableFlags::PRESENT | PageTableFlags::WRITABLE,
-			);
-			pdpte[0].set_addr(BOOT_PDE, PageTableFlags::PRESENT | PageTableFlags::WRITABLE);
-
-			for i in 0..512 {
-				let addr = PhysAddr::new(i as u64 * Page::<Size2MiB>::SIZE);
-				pde[i].set_addr(
-					addr,
-					PageTableFlags::PRESENT | PageTableFlags::WRITABLE | PageTableFlags::HUGE_PAGE,
-				);
-			}
-		}
+	fn init_guest_mem(&mut self) {
+		self.mem.init_guest_mem();
 	}
 }
 
@@ -293,68 +244,3 @@ impl Vm for Uhyve {
 #[allow(clippy::non_send_fields_in_send_ty)]
 unsafe impl Send for Uhyve {}
 unsafe impl Sync for Uhyve {}
-
-#[derive(Debug)]
-struct MmapMemory {
-	flags: u32,
-	memory_size: usize,
-	guest_address: usize,
-	host_address: usize,
-}
-
-impl MmapMemory {
-	pub fn new(
-		flags: u32,
-		memory_size: usize,
-		guest_address: u64,
-		huge_pages: bool,
-		mergeable: bool,
-	) -> MmapMemory {
-		let host_address = unsafe {
-			mmap_anonymous(
-				None,
-				memory_size.try_into().unwrap(),
-				ProtFlags::PROT_READ | ProtFlags::PROT_WRITE,
-				MapFlags::MAP_PRIVATE | MapFlags::MAP_NORESERVE,
-			)
-			.expect("mmap failed")
-		};
-
-		if mergeable {
-			debug!("Enable kernel feature to merge same pages");
-			unsafe {
-				madvise(host_address, memory_size, MmapAdvise::MADV_MERGEABLE).unwrap();
-			}
-		}
-
-		if huge_pages {
-			debug!("Uhyve uses huge pages");
-			unsafe {
-				madvise(host_address, memory_size, MmapAdvise::MADV_HUGEPAGE).unwrap();
-			}
-		}
-
-		MmapMemory {
-			flags,
-			memory_size,
-			guest_address: guest_address as usize,
-			host_address: host_address.as_ptr() as usize,
-		}
-	}
-
-	#[allow(dead_code)]
-	fn as_slice_mut(&mut self) -> &mut [u8] {
-		unsafe { std::slice::from_raw_parts_mut(self.host_address as *mut u8, self.memory_size) }
-	}
-}
-
-impl Drop for MmapMemory {
-	fn drop(&mut self) {
-		if self.memory_size > 0 {
-			let host_addr = NonNull::new(self.host_address as *mut c_void).unwrap();
-			unsafe {
-				munmap(host_addr, self.memory_size).unwrap();
-			}
-		}
-	}
-}
