+++ src/macos/aarch64/uhyve.rs
@@ -160,7 +160,7 @@ impl Vm for Uhyve {
 		self.boot_info = header;
 	}
 
-	fn init_guest_mem(&self) {
+	fn init_guest_mem(&mut self) {
 		debug!("Initialize guest memory");
 
 		let (mem_addr, _) = self.guest_mem();
